pyproject.toml
[project]
name = "pretty-mod"
description = "A python module tree explorer for LLMs (and humans)"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "zzstoatzz", email = "thrast36@gmail.com" }]
requires-python = ">=3.9"
dynamic = ["version"]
dependencies = []

[project.scripts]
pretty-mod = "pretty_mod.cli:main"

[dependency-groups]
dev = [
    "maturin",
    "pytest-sugar",
    "ruff",
    "ty",
    "ipython",
    "pre-commit",
    "pyinstrument",
    "rich",
]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "pretty_mod._pretty_mod"
bindings = "pyo3"

[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[tool.ruff.lint]
extend-select = ["I", "UP"]

[tool.ty.src]
root = "./src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--strict-markers"]
